prepared base interfaces and implemented saving trainee

+++ src/main/resources/db/changelog/tables/trainee.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS trainee (
     id BIGSERIAL PRIMARY KEY,
-    user_id INT NOT NULL UNIQUE,
+    user_id BIGSERIAL NOT NULL UNIQUE,
     date_of_birth DATE,
     address VARCHAR(256),
